Added "github-actions-rest.yaml" and updated rest-api tests
After merging PR #3238 (job Engine as standalone service), the rest-api tests
PR/branch was not working anymore, becasue one of the steps was to pull the
Kapua docker images from Docker Huh and deploy them (job Engine image is not yet
on the Docker Hub).
Because of this we decided to sepparate rest-api tests from all others (unit and
integration tests), which now includes building Kapua with -Pdocker command,
which will build the docker images.
Afterwards we deploy these images with ./docker-deploy.sh command (this command
will not look for local images and deply them and not download the images from
Docker Hub).

After the rest-api tests are run, we undpeloy docker containers.

I have deleted the rest-api stage in the "github-actions.yaml" file, as
it is not needed anymore.

Also the "codecov" part in the "build" stage has been deleted, becasue there
is no point in running code coverage after the build.

Signed-off-by: Leonardo Gaube <leonardo.gaube@endava.com>